requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
transformers
torch
opencv-python
